Add new tests and logger for debuging
Testcases with bigger structure still fail. Further test has been added
that checks for the determinant of the transformation matrix. Now some
lines of code have been added to fulfill this test. But now enough tests
have been added to test different implementations of the get_transform
function.

Things that may be responsible for the bugs:
        - QR factorization does not consistently yield a basis with same
          orientation ( maybe its not intended to do so and dialanine
          was just luck)
        - -1 to ensure that HA is behind CA and the rest might cause
          issues, now replaced by rotation matrix multiplication.

Things to try instead:
        - rotation matrices
        - quarternions