Update clashes implementation and coloring
The coloring for Terminal and PyMOL display is updated to match the
publication. Also, the arangement to prioritize coloring severe things
last has been adjusted.

In the clashes calculation, previously only the VDW radius of one
residue was considered. The code now properly calculates the clashes
considering both VDW radii.
All pair of atoms within 5 Angstroms of each other are checked for
clashes. With a higher cutoff, the number of pairs to consider is too
high, leading to a lot longer computation time.
Pairs where, atom a and b are in the same residue are ignored. Further,
If atom a and b are in residues next to each other, and both atoms are
part of the backbone, the pair is ignored.
If a clash between residue a and residue b already has been calculated,
ignore the pair.

CD is added to the backbone, which is specific to prolines.
Cysteine disulfide bridges are ignored for clashes. as they are
covalently linked.

All pairs, between which the distance is lower than their combined vdw
radii - 0.5, are considered a clash.
The -0.5 is chosen arbitrarily.